Informando o id_user nas operações do crud

+++ p4-item-catalog/project.py
@@ -220,6 +220,7 @@ def showRestaurants():
 # Create new restaurant
 
 
+'''
 @app.route('/restaurants/new/', methods=['GET', 'POST'])
 def newRestaurant():
     if 'username' not in login_session:
@@ -231,10 +232,26 @@ def newRestaurant():
         return redirect(url_for('showRestaurants'))
     else:
         return render_template('newRestaurant.html')
+'''
+
+
+@app.route('/restaurants/new/', methods=['GET', 'POST'])
+def newRestaurant():
+    if 'username' not in login_session:
+        return redirect('/login')
+    if request.method == 'POST':
+        newRestaurant = Restaurant(
+            name=request.form['name'],
+            user_id=login_session['user_id'])
+        session.add(newRestaurant)
+        flash('New Restaurant %s Successfully Created' % newRestaurant.name)
+        session.commit()
+        return redirect(url_for('showRestaurants'))
+    else:
+        return render_template('newRestaurant.html')
+
 
 # Edit restaurant
-
-
 @app.route('/restaurants/<int:restaurant_id>/edit/', methods=['GET', 'POST'])
 def editRestaurant(restaurant_id):
     editedRestaurant = session.query(
@@ -246,6 +263,7 @@ def editRestaurant(restaurant_id):
     if request.method == 'POST':
         if request.form['name']:
             editedRestaurant.name = request.form['name']
+            editedRestaurant.user_id = login_session['user_id']
             flash('Restaurante editado com sucesso %s' % editedRestaurant.name)
             return redirect(url_for('showRestaurants'))
     else:
@@ -298,7 +316,9 @@ def newMenuItem(restaurant_id):
         newItem = MenuItem(name=request.form['name'],
                            description=request.form['description'],
                            price=request.form['price'],
-                           restaurant_id=restaurant_id)
+                           restaurant_id=restaurant_id,
+                           # user_id=restaurant.user_id
+                           user_id=login_session['user_id'])
         session.add(newItem)
         session.commit()
         flash("Item de menu foi adicionado")
@@ -307,7 +327,7 @@ def newMenuItem(restaurant_id):
         return render_template('newmenuitem.html', restaurant_id=restaurant_id)
     return render_template('newMenuItem.html', restaurant=restaurant)
 
-# Editar menu item
+# Edit menu item
 
 
 @app.route('/restaurants/<int:restaurant_id>/<int:menu_id>/edit/',
@@ -316,6 +336,9 @@ def editMenuItem(restaurant_id, menu_id):
     if 'username' not in login_session:
         return redirect('/login')
     editedItem = session.query(MenuItem).filter_by(id=menu_id).one()
+    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    if login_session['user_id'] != restaurant.user_id:
+        return "<script>{alert('Unauthorized');}</script>"
     if request.method == 'POST':
         if request.form['name']:
             editedItem.name = request.form['name']
@@ -323,17 +346,19 @@ def editMenuItem(restaurant_id, menu_id):
             editedItem.description = request.form['description']
         if request.form['price']:
             editedItem.price = request.form['price']
+        if request.form['course']:
+            editedItem.course = request.form['course']
         session.add(editedItem)
         session.commit()
         flash("Item de menu foi editado")
-
         return redirect(url_for('restaurantMenu', restaurant_id=restaurant_id))
     else:
-        return render_template(
-            'editmenuitem.html', restaurant_id=restaurant_id,
-            menu_id=menu_id, item=editedItem)
+        return render_template('editmenuitem.html',
+                               restaurant_id=restaurant_id,
+                               menu_id=menu_id,
+                               item=editedItem)
 
-# Deletar menu item
+# Delet menu item
 
 
 @app.route('/restaurants/<int:restaurant_id>/<int:menu_id>/delete/',
@@ -341,7 +366,10 @@ def editMenuItem(restaurant_id, menu_id):
 def deleteMenuItem(restaurant_id, menu_id):
     if 'username' not in login_session:
         return redirect('/login')
+    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
     itemToDelete = session.query(MenuItem).filter_by(id=menu_id).one()
+    if login_session['user_id'] != restaurant.user_id:
+        return "<script>{alert('Unauthorized');}</script>"
     if request.method == 'POST':
         session.delete(itemToDelete)
         session.commit()
